+++ rpc/getSlotLeaders.go
@@ -9,8 +9,8 @@ import (
 // GetSlotLeaders returns the slot leaders for a given slot range.
 func (cl *Client) GetSlotLeaders(
 	ctx context.Context,
-	start int,
-	limit int,
+	start uint64,
+	limit uint64,
 ) (out []solana.PublicKey, err error) {
 	params := []interface{}{start, limit}
 	err = cl.rpcClient.CallFor(&out, "getSlotLeaders", params)
